Extract linodeEvents middleware.
